+++ web/documentserver-example/nodejs/helpers/documentService.js
@@ -61,9 +61,11 @@ documentService.getConvertedUri = function getConvertedUri(
   filePass = null,
   lang = null,
 ) {
-  const fromExt = fromExtension || fileUtility.getFileExtension(documentUri); // get the current document extension
+  // get the current document extension
+  const fromExt = fromExtension || fileUtility.getFileExtension(documentUri, false, true);
 
-  const title = fileUtility.getFileName(documentUri) || guidManager.newGuid(); // get the current document name or uuid
+  // get the current document name or uuid
+  const title = fileUtility.getFileNameFromUrl(documentUri) || guidManager.newGuid();
 
   // generate the document key value
   const revisionId = documentService.generateRevisionId(documentRevisionId || documentUri);