+++ web/documentserver-example/ruby/app/configuration/configuration_tests.rb
@@ -151,7 +151,7 @@ class ConfigurationManagerDocumentServerCommandURITests < Test::Unit::TestCase
     uri = config_manager.document_server_command_uri
     assert_equal(
       uri.to_s,
-      'http://documentserver/coauthoring/CommandService.ashx'
+      'http://documentserver/command'
     )
   end
 
@@ -175,7 +175,7 @@ class ConfigurationManagerDocumentServerConverterURITests < Test::Unit::TestCase
     uri = config_manager.document_server_converter_uri
     assert_equal(
       uri.to_s,
-      'http://documentserver/ConvertService.ashx'
+      'http://documentserver/converter'
     )
   end