+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+- he-IL skin language
+- ConvertService.ashx address replaced with converter
+- coauthoring/CommandService.ashx address replaced with command
+
 - golang: new integration example
 - golang: upload files to the server
 - golang: create blank files and files with sample content